requirements.txt
pandas>=1.3.3
numpy>=1.21.2
scikit-learn>=1.0.0
torch>=1.9.0
torchvision>=0.10.0 # Often included with torch, but good to specify
torchaudio>=0.9.0    # Often included with torch, but good to specify
wandb>=0.15.0
pyyaml>=6.0
matplotlib>=3.4.3
seaborn>=0.11.2
jupyter>=1.0.0 # For the notebooks
imbalanced-learn>=0.9.0 
wandb